ramips: add support for the TP-Link TL-WA850RE v4
TP-Link TL-WA850RE is a wireless range extender with Ethernet and 2.4G
WiFi with internal antennas. It's based on a MediaTek MT7628NN.

Specifications
--------------

- MediaTek MT7628NN (575 Mhz)
- 32 MB of RAM
- 4 MB of FLASH (SPI)
- 2T2R 2.4 Ghz WiFi
- 1x 10/100 Mbps Ethernet
- UART header on PCB (57600 8n1)
- 9x LED (GPIO), 2x button

Installation
------------

Web Interface
-------------

Installing via the web interface works by sending the OpenWrt factory
image as is. Be warned that there is no other way to recover except
to open the case in case the install/upgrade fails.

Serial console
--------------

Opening the case is quite hard, since it is welded together. Rename the
OpenWrt sysupgrade image to "test.bin", then plug in the device and
quickly press "2" to enter flash mode (no line feed).
Follow the prompts until OpenWrt is installed.

Unfortunately, this device does not offer a recovery mode or a tftp
installation method without using the serial port. If the web interface
upgrade fails, you have to open your device and attach serial console.
It is also possible to brick the device if something goes wrong,
as the bootloader is not write protected.

Additional notes
----------------

MAC address assignment is based on stock-firmware. For me, the device
assigns the MAC on the label to Ethernet and +1 of that to the 2.4G WiFi.

Signed-off-by: Urja Rannikko <urjaman@gmail.com>